Consumo completo com CRUD e buscas

+++ blogPessoal/src/app/delete-tema/delete-tema.component.ts
@@ -1,3 +1,4 @@
+import { AlertasService } from './../service/alertas.service';
 import { Router, ActivatedRoute } from '@angular/router';
 import { TemaService } from './../service/tema.service';
 import { Tema } from './../model/Tema';
@@ -16,7 +17,7 @@ export class DeleteTemaComponent implements OnInit {
     private temaService: TemaService,
     private router: Router,
     private route: ActivatedRoute,
-   
+    private alerta: AlertasService
   ) { }
 
   ngOnInit() {
@@ -33,10 +34,16 @@ export class DeleteTemaComponent implements OnInit {
   }
 
   btnSim() {
-    this.temaService.deleteTema(this.tema.id).subscribe(() => {
+    if (this.tema.postagem.length != 0) {
+      this.alerta.showAlertDanger('Esse tema não pode ser modificado, pois já pertence a uma postagem.')
       this.router.navigate(['/cadastro-tema'])
-      alert('Tema apagado com sucesso!')
-    })
+    } else {
+      this.temaService.deleteTema(this.tema.id).subscribe(() => {
+        this.router.navigate(['/cadastro-tema'])
+        this.alerta.showAlertSuccess('Tema apagado com sucesso!')
+      })
+    }
+
   }
 
   btnNao() {
